Promote flux gradient to double

+++ src/chemistry/GEAR_MF_DIFFUSION/chemistry_gradients.h
@@ -154,7 +154,7 @@ __attribute__((always_inline)) INLINE static void chemistry_gradients_collect(
 
     // TODO: move this to a function inside hyperbolic/gradients
 #if defined(CHEMISTRY_GEAR_MF_HYPERBOLIC_DIFFUSION)
-    const float dflux[3] = {
+    const double dflux[3] = {
         chi->flux[g][0] - chj->flux[g][0],
         chi->flux[g][1] - chj->flux[g][1],
         chi->flux[g][2] - chj->flux[g][2]};
@@ -351,7 +351,7 @@ chemistry_gradients_nonsym_collect(float r2, const float *dx, float hi,
     chemistry_part_update_metal_mass_fraction_gradients(pi, g, dU_i);
 
 #if defined(CHEMISTRY_GEAR_MF_HYPERBOLIC_DIFFUSION)
-    const float dflux[3] = {
+    const double dflux[3] = {
         chi->flux[g][0] - chj->flux[g][0],
         chi->flux[g][1] - chj->flux[g][1],
         chi->flux[g][2] - chj->flux[g][2]};
